+++ metadata.php
@@ -570,5 +570,11 @@ $aModule = [
             'value' => 3.5,
             'group' => null
         ],
+        [
+            'name' => 'oscPayPalUseStructuralCustomIdSchema',
+            'type' => 'bool',
+            'value' => false,
+            'group' => null
+        ],
     ],
 ];